# docs/api/openapi-etapa1.yaml
openapi: 3.1.0
info:
  title: Cerniq.app API — Etapa 1 (Data Enrichment)
  version: 1.0.0
  description: |
    Specificație OpenAPI pentru Etapa 1 - Data Enrichment Pipeline.

    ## Overview
    Etapa 1 implementează pipeline-ul de îmbogățire date pentru prospecți B2B:
    - **Bronze Layer**: Ingestie și validare date brute (CSV, Excel, API, Webhook)
    - **Silver Layer**: Îmbogățire date (ANAF, Termene.ro, ONRC, Geocoding)
    - **Gold Layer**: Lead-uri calificate pentru outreach
    - **HITL**: Human-in-the-Loop pentru aprobare date incerte

    ## Autentificare
    Toate endpoint-urile necesită autentificare JWT:
    ```
    Authorization: Bearer <access_token>
    X-Tenant-ID: <tenant_uuid>
    ```

    ## Rate Limiting
    - Standard: 100 req/min
    - Import: 10 req/min
    - Enrichment: 60 req/min

    Pentru schema globală și convenții comune, vezi docs/api/openapi.yaml.
  contact:
    name: Cerniq Development Team
    email: dev@cerniq.app
  license:
    name: Proprietary
    url: https://cerniq.app/license
servers:
  - url: https://api.cerniq.app/api/v1
    description: Production
  - url: https://staging.cerniq.app/api/v1
    description: Staging
  - url: http://localhost:64000/api/v1
    description: Development
...
paths:
...
  /dashboard/stats:
    get:
      tags: [Dashboard]
      operationId: getDashboardStats
      summary: Obține statistici generale dashboard
      description: |
        Returnează statistici agregate pentru toate layerele:
        - Bronze: total, pending, promoted, rejected
        - Silver: validated, enriched, eligible
        - Gold: promoted leads
        - Approvals: pending, urgent
        - Quality: average score, distribution
        - Enrichment: queue depths
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Statistici dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStatsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dashboard/activity:
    get:
      tags: [Dashboard]
      operationId: getDashboardActivity
      summary: Obține activitatea recentă
      description: Feed de activitate cu evenimente recente (import-uri, aprobări, erori)
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Număr maxim de activități returnate
      responses:
        "200":
          description: Lista de activități
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"

  # ==================== IMPORTS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obținut din /api/v1/auth/login
...
  responses:
    Unauthorized:
      description: Autentificare necesară sau token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    DashboardStatsResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            bronze:
              type: object
              properties:
                total:
                  type: integer
                today:
                  type: integer
                pending:
                  type: integer
                promoted:
                  type: integer
                rejected:
                  type: integer
            silver:
              type: object
              properties:
                total:
                  type: integer
                validated:
                  type: integer
                enriched:
                  type: integer
                eligible:
                  type: integer
            gold:
              type: object
              properties:
                total:
                  type: integer
                promoted:
                  type: integer
                thisWeek:
                  type: integer
            approvals:
              type: object
              properties:
                pending:
                  type: integer
                urgent:
                  type: integer
                avgResolutionHours:
                  type: number
            quality:
              type: object
              properties:
                average:
                  type: number
                distribution:
                  type: object
                  properties:
                    high:
                      type: integer
                    medium:
                      type: integer
                    low:
                      type: integer
            enrichment:
              type: object
              properties:
                queues:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      depth:
                        type: integer
                      processing:
                        type: integer
                completionRate:
                  type: number
    ActivityListResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              type:
                type: string
              title:
                type: string
              description:
                type: string
              timestamp:
                type: string
                format: date-time
              metadata:
                type: object

    # ========== IMPORTS ==========
...
security:
  - BearerAuth: []